requirements.txt
redis==3.2.0
Django==2.2.6
celery==4.4.0
eventlet==0.26.1
django-celery-beat==2.0.0
django-celery-results==1.2.1
django-cors-headers==3.5.0

mysqlclient==1.4.4
MarkupSafe==1.1.1
Mako==1.0.6
requests==2.22.0
python-json-logger==0.1.7
whitenoise==3.3.0
six==1.11.0

cx-Oracle==8.1.0
django-haystack==3.0
elasticsearch==5.5.3